test: Move parameters initialization test to platform scope
