+++ tests/test_parameters.py
@@ -6,8 +6,7 @@ from qibolab._core.components.configs import Config
 from qibolab._core.native import Native, TwoQubitNatives
 from qibolab._core.parameters import ConfigKinds, Parameters, TwoQubitContainer
 from qibolab._core.platform.load import create_platform
-from qibolab._core.pulses.pulse import Pulse, Readout
-from qibolab.platform import Hardware, initialize_parameters
+from qibolab._core.pulses.pulse import Pulse
 
 
 def test_two_qubit_container():
@@ -38,9 +37,9 @@ class DummyConfig1(Config):
 
 
 class TestConfigKinds:
-    # TODO: add @staticmethod and drop unused `self`, once py3.9 will be abandoned
+    @staticmethod
     @pytest.fixture(autouse=True)
-    def clean_kinds(self):
+    def clean_kinds():
         ConfigKinds.reset()
 
     def test_manipulation(self):
@@ -105,37 +104,3 @@ def test_update():
     assert dummy.settings.nshots == 42
     assert dummy.natives.single_qubit[1].RX[0][1].amplitude == -0.123
     assert dummy.natives.single_qubit[1].RX[0][1].duration == 456.7
-
-
-def test_builder():
-    dummy = create_platform("dummy")
-
-    hardware = Hardware(
-        instruments=dummy.instruments,
-        qubits=dummy.qubits,
-        couplers=dummy.couplers,
-    )
-    parameters = initialize_parameters(
-        hardware=hardware, natives=["RX", "MZ", "CZ"], pairs=["0-2"]
-    )
-
-    for q in dummy.qubits:
-        assert f"{q}/drive" in parameters.configs
-        assert f"{q}/probe" in parameters.configs
-        assert f"{q}/acquisition" in parameters.configs
-        assert f"{q}/drive12" in parameters.configs
-        assert q in parameters.native_gates.single_qubit
-    for c in dummy.couplers:
-        assert f"coupler_{c}/flux" in parameters.configs
-        assert c in parameters.native_gates.coupler
-
-    assert list(parameters.native_gates.two_qubit) == [(0, 2)]
-    sequence = parameters.native_gates.two_qubit[(0, 2)].CZ
-    assert sequence[0][0] == "0/flux"
-    assert isinstance(sequence[0][1], Pulse)
-    sequence = parameters.native_gates.single_qubit[0].RX
-    assert sequence[0][0] == "0/drive"
-    assert isinstance(sequence[0][1], Pulse)
-    sequence = parameters.native_gates.single_qubit[2].MZ
-    assert sequence[0][0] == "2/acquisition"
-    assert isinstance(sequence[0][1], Readout)
